Added Navigation Component and also Setup React-Router in Project
