+++ crwn-clothing/.gitignore
@@ -5,6 +5,9 @@
 /.pnp
 .pnp.js
 
+package-lock.json
+package.json
+
 # testing
 /coverage
