+++ crwn-clothing/src/index.scss
@@ -2,3 +2,11 @@ body {
     margin: 0;
     font-family: 'Open Sans', sans-serif;
 }
+
+a{
+    text-decoration: none;
+    color: black;
+}
+*{
+    box-sizing: border-box;
+}
